pyproject.toml
[project]
name = "arch-design-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.1.1",
    "httpx>=0.28.1",
    "msal>=1.31.0",
    "pymupdf>=1.25.0",
    "pypdf>=6.9.0",
    "python-dotenv>=1.0.0",
]
